Update the Encoders
