+++ microbat/src/main/microbat/baseline/constraints/VariableConstraintA1.java
@@ -1,6 +1,6 @@
 package microbat.baseline.constraints;
 
-import microbat.baseline.encoders.PropabilityInference;
+import microbat.baseline.beliefpropagation.PropabilityInference;
 import microbat.model.trace.TraceNode;
 import microbat.model.value.VarValue;
